Fix user add to use onboarding flow
