+++ backend/cmd/server/bootstrap/01-default-resources.yaml
@@ -1603,14 +1603,10 @@ nodes:
     components:
     - align: left
       type: TEXT
-      id: admin_registration_complete_icon
-      label: ✅
-      variant: HEADING_1
-    - align: left
-      type: TEXT
       id: admin_registration_complete_heading
       label: '{{ t(onboarding:forms.add_user_complete.title) }}'
       variant: HEADING_1
+      color: SUCCESS
     - align: left
       type: TEXT
       id: admin_registration_complete_message
@@ -1762,14 +1758,10 @@ nodes:
     components:
     - align: left
       type: TEXT
-      id: email_status_icon
-      label: ✅
-      variant: HEADING_1
-    - align: left
-      type: TEXT
       id: email_status_heading
       label: '{{ t(onboarding:forms.invite_email_sent.title) }}'
       variant: HEADING_1
+      color: SUCCESS
     - align: left
       type: TEXT
       id: email_status_message
@@ -1790,14 +1782,10 @@ nodes:
     components:
     - align: left
       type: TEXT
-      id: email_unavailable_icon
-      label: ⚠️
-      variant: HEADING_1
-    - align: left
-      type: TEXT
       id: email_unavailable_heading
       label: '{{ t(onboarding:forms.invite_email_unavailable.title) }}'
       variant: HEADING_1
+      color: ERROR
     - align: left
       type: TEXT
       id: email_unavailable_message
@@ -1833,11 +1821,6 @@ nodes:
   type: PROMPT
   meta:
     components:
-    - align: left
-      type: TEXT
-      id: link_status_icon
-      label: ✅
-      variant: HEADING_1
     - align: left
       type: TEXT
       id: link_status_heading
@@ -1937,14 +1920,10 @@ nodes:
     components:
     - align: left
       type: TEXT
-      id: registration_complete_icon
-      label: ✅
-      variant: HEADING_1
-    - align: left
-      type: TEXT
       id: registration_complete_heading
       label: '{{ t(invite:complete.title) }}'
       variant: HEADING_1
+      color: SUCCESS
     - align: left
       type: TEXT
       id: registration_complete_message
