+++ docs/src/theme/MDXComponents.tsx
@@ -17,12 +17,27 @@
  */
 
 import MDXComponents from '@theme-original/MDXComponents';
-import {Box, Card, CardContent, Typography, ColorSchemeSVG, Table, TableBody, TableCell, TableContainer, TableHead, TableRow} from '@wso2/oxygen-ui';
+import {
+  Box,
+  Card,
+  CardContent,
+  Typography,
+  ColorSchemeSVG,
+  Table,
+  TableBody,
+  TableCell,
+  TableContainer,
+  TableHead,
+  TableRow,
+} from '@wso2/oxygen-ui';
 import {AgentInteractionsDiagram} from '@site/src/components/AgentInteractionsDiagram';
 import {LangTabs, Lang} from '@site/src/components/AgentLang';
 import {AgentModeSelector, Mode} from '@site/src/components/AgentMode';
 import {AgentOwnTokenFlow, AgentOboFlow} from '@site/src/components/AgentQuickstartFlow';
-import {AIAgentIdentityRoadmap as AIAgentIdentityExplorer, AIAgentSolutionPatternsRoadmap} from '@site/src/components/AIAgentIdentityJourney';
+import {
+  AIAgentIdentityRoadmap as AIAgentIdentityExplorer,
+  AIAgentSolutionPatternsRoadmap,
+} from '@site/src/components/AIAgentIdentityJourney';
 import ApiReference from '@site/src/components/ApiReference';
 import ApiVersionReference from '@site/src/components/ApiVersionReference';
 import {
