+++ frontend/apps/console/src/features/trusted-issuers/utils/mapConnectionToTrustedIssuer.ts
@@ -16,8 +16,8 @@
  * under the License.
  */
 
-import type {TrustedIssuer} from '../models/trusted-issuer';
 import type {ConnectionResponse} from '@thunderid/configure-connections';
+import type {TrustedIssuer} from '../models/trusted-issuer';
 
 /**
  * Maps an OIDC connection API response to the narrower trusted-issuer shape this feature works
