+++ frontend/packages/configure-users/src/index.ts
@@ -60,7 +60,6 @@ export * from './models/users';
 export {default as UserAddPage} from './pages/UserAddPage';
 export {default as UserCreatePage} from './pages/UserCreatePage';
 export {default as UserEditPage} from './pages/UserEditPage';
-export {default as UserInvitePage} from './pages/UserInvitePage';
 export {default as UsersListPage} from './pages/UsersListPage';
 
 // Routes
