+++ frontend/packages/configure-users/src/pages/UserCreatePage.tsx
@@ -16,384 +16,279 @@
  * under the License.
  */
 
-import {useGetChildOrganizationUnits} from '@thunderid/configure-organization-units';
+import {OrganizationUnitTreePicker} from '@thunderid/configure-organization-units';
 import {useLogger} from '@thunderid/logger/react';
-import {useThunderID} from '@thunderid/react';
+import {
+  EmbeddedFlowComponentType,
+  InviteUser,
+  useThunderID,
+  type EmbeddedFlowComponent,
+  type InviteUserRenderProps,
+} from '@thunderid/react';
 import {
   Box,
   Stack,
+  Typography,
   Button,
+  TextField,
   IconButton,
+  FormControl,
+  FormLabel,
+  Select,
+  MenuItem,
   LinearProgress,
   AppBreadcrumbs,
-  Typography,
-  Alert,
-  Snackbar,
 } from '@wso2/oxygen-ui';
 import {X} from '@wso2/oxygen-ui-icons-react';
 import type {JSX} from 'react';
-import {useState, useCallback, useMemo, useEffect} from 'react';
+import {useState, useCallback, useEffect, useRef} from 'react';
 import {useTranslation} from 'react-i18next';
 import {useNavigate} from 'react-router';
-import useCreateUser from '../api/useCreateUser';
-import useGetUserType from '../api/useGetUserType';
-import useGetUserTypes from '../api/useGetUserTypes';
-import ConfigureOrganizationUnit from '../components/create-user/ConfigureOrganizationUnit';
-import ConfigureUserDetails from '../components/create-user/ConfigureUserDetails';
-import ConfigureUserType from '../components/create-user/ConfigureUserType';
-import useUserCreate from '../contexts/UserCreate/useUserCreate';
 import useUserRoutes from '../hooks/useUserRoutes';
-import {UserCreateFlowStep} from '../models/user-create-flow';
+
+const ONBOARDING_MODE_CREATE_ACTION_ID = 'action_create_user_now';
+
+/** Typed shape for flow sub-components */
+type FlowSubComponent = EmbeddedFlowComponent & {
+  components?: EmbeddedFlowComponent[];
+  hint?: string;
+  options?: unknown[];
+  placeholder?: string;
+  required?: boolean;
+  variant?: string;
+};
+
+function UserCreateStepContent({
+  renderProps,
+  onStepLabelChange,
+}: {
+  renderProps: InviteUserRenderProps;
+  onStepLabelChange: (label: string) => void;
+}): JSX.Element {
+  const {resolveFlowTemplateLiterals: rawResolve} = useThunderID();
+  const resolve = useCallback((text?: string) => (text ? rawResolve(text) : undefined), [rawResolve]);
+  const {t} = useTranslation();
+  const components = renderProps.components as EmbeddedFlowComponent[] | undefined;
+
+  // Auto-select create action if present
+  const createAction = components?.find((c) => c.id === ONBOARDING_MODE_CREATE_ACTION_ID);
+
+  const submittedActionIdRef = useRef<string | undefined>(undefined);
+
+  useEffect(() => {
+    if (createAction && submittedActionIdRef.current !== createAction.id) {
+      submittedActionIdRef.current = createAction.id;
+      renderProps.handleSubmit(createAction, renderProps.values).catch(() => undefined);
+    }
+  }, [createAction, renderProps]);
+
+  // Derive step label from HEADING_1 component
+  const stepLabel = components?.find(
+    (c) =>
+      (String(c.type) === String(EmbeddedFlowComponentType.Text) || c.type === 'TEXT') &&
+      (c as FlowSubComponent).variant === 'HEADING_1',
+  );
+
+  useEffect(() => {
+    if (stepLabel && typeof stepLabel.label === 'string') {
+      onStepLabelChange(t(resolve(stepLabel.label) ?? stepLabel.label));
+    }
+  }, [stepLabel, resolve, t, onStepLabelChange]);
+
+  const {values, additionalData, handleInputChange} = renderProps;
+
+  const renderComponent = (component: EmbeddedFlowComponent, index: number): JSX.Element | null => {
+    // Render text components
+    if (String(component.type) === String(EmbeddedFlowComponentType.Text) || component.type === 'TEXT') {
+      const label = typeof component.label === 'string' ? component.label : '';
+      const variant = (component as FlowSubComponent).variant;
+
+      if (variant === 'HEADING_1') {
+        return (
+          <Typography key={component.id ?? index} variant="h1" gutterBottom>
+            {t(resolve(label) ?? label)}
+          </Typography>
+        );
+      }
+
+      return (
+        <Typography key={component.id ?? index} variant="body1" color="text.secondary">
+          {t(resolve(label) ?? label)}
+        </Typography>
+      );
+    }
+
+    // Render form fields
+    if (
+      component.type === 'EMAIL_INPUT' ||
+      component.type === 'TEXT_INPUT' ||
+      component.type === 'PHONE_INPUT' ||
+      component.type === 'PASSWORD_INPUT'
+    ) {
+      const ref = component.ref;
+      const label = typeof component.label === 'string' ? component.label : '';
+      const placeholder = (component as FlowSubComponent).placeholder ?? '';
+      const required = (component as FlowSubComponent).required ?? false;
+
+      if (!ref) return null;
+
+      const value = ((values as Record<string, unknown>)?.[ref] as string) ?? '';
+      let inputType = 'text';
+
+      if (component.type === 'EMAIL_INPUT') inputType = 'email';
+      if (component.type === 'PASSWORD_INPUT') inputType = 'password';
+      if (component.type === 'PHONE_INPUT') inputType = 'tel';
+
+      return (
+        <FormControl key={component.id ?? index} fullWidth required={required}>
+          <FormLabel htmlFor={ref}>{t(resolve(label) ?? label)}</FormLabel>
+          <TextField
+            id={ref}
+            type={inputType}
+            size="small"
+            variant="outlined"
+            placeholder={resolve(placeholder) ?? placeholder}
+            value={value}
+            required={required}
+            onChange={(e) => handleInputChange(ref, e.target.value)}
+          />
+        </FormControl>
+      );
+    }
+
+    // Render SELECT
+    if (component.type === 'SELECT') {
+      const ref = component.ref;
+      const label = typeof component.label === 'string' ? component.label : '';
+      const options = (component as FlowSubComponent).options ?? [];
+      const required = (component as FlowSubComponent).required ?? false;
+
+      if (!ref || !options.length) return null;
+
+      const value = ((values as Record<string, unknown>)?.[ref] as string) ?? '';
+
+      return (
+        <FormControl key={component.id ?? index} fullWidth required={required}>
+          <FormLabel htmlFor={ref}>{t(resolve(label) ?? label)}</FormLabel>
+          <Select
+            id={ref}
+            value={value}
+            size="small"
+            displayEmpty
+            required={required}
+            onChange={(e) => handleInputChange(ref, String(e.target.value))}
+          >
+            <MenuItem value="">{t('Select an option')}</MenuItem>
+            {options.map((opt: unknown) => {
+              const optValue =
+                typeof opt === 'object' && opt !== null ? (opt as Record<string, unknown>)['value'] : opt;
+              const optLabel =
+                typeof opt === 'object' && opt !== null ? (opt as Record<string, unknown>)['label'] : opt;
+              const keyValue = String(optValue);
+              return (
+                <MenuItem key={keyValue} value={keyValue}>
+                  {String(optLabel)}
+                </MenuItem>
+              );
+            })}
+          </Select>
+        </FormControl>
+      );
+    }
+
+    // Render OU_SELECT
+    if (component.type === 'OU_SELECT') {
+      const ref = component.ref;
+      const label = typeof component.label === 'string' ? component.label : '';
+      const required = (component as FlowSubComponent).required ?? false;
+
+      if (!ref) return null;
+
+      const value = ((values as Record<string, unknown>)?.[ref] as string) ?? '';
+      const rootOuId = additionalData?.['rootOuId'] as string | undefined;
+
+      return (
+        <FormControl key={component.id ?? index} fullWidth required={required}>
+          <FormLabel htmlFor={ref}>{t(resolve(label) ?? label)}</FormLabel>
+          <OrganizationUnitTreePicker
+            value={value}
+            onChange={(ouId: string) => handleInputChange(ref, ouId)}
+            rootOuId={rootOuId}
+          />
+        </FormControl>
+      );
+    }
+
+    // Render ACTION buttons (submit buttons)
+    if (component.type === 'ACTION') {
+      const label = typeof component.label === 'string' ? component.label : '';
+      const variant = (component as FlowSubComponent).variant;
+
+      return (
+        <Button
+          key={component.id ?? index}
+          variant={variant === 'OUTLINED' ? 'outlined' : 'contained'}
+          onClick={() => {
+            renderProps.handleSubmit(component, renderProps.values).catch(() => undefined);
+          }}
+          fullWidth
+          size="large"
+        >
+          {t(resolve(label) ?? label)}
+        </Button>
+      );
+    }
+
+    // Render nested components (BLOCK, STACK)
+    if ((component as FlowSubComponent).components) {
+      const nested = (component as FlowSubComponent).components!;
+      return (
+        <Stack key={component.id ?? index} direction="column" spacing={2}>
+          {nested.map((c, idx) => renderComponent(c, idx))}
+        </Stack>
+      );
+    }
+
+    return null;
+  };
+
+  return (
+    <Stack direction="column" spacing={4}>
+      {components?.map((component: EmbeddedFlowComponent, index: number) => renderComponent(component, index))}
+    </Stack>
+  );
+}
 
 export default function UserCreatePage(): JSX.Element {
   const {t} = useTranslation();
   const navigate = useNavigate();
-  const logger = useLogger('UserCreatePage');
-  const createUserMutation = useCreateUser();
   const routes = useUserRoutes();
+  const logger = useLogger('UserCreatePage');
+  const [breadcrumbs, setBreadcrumbs] = useState<string[]>([]);
 
-  const {
-    currentStep,
-    setCurrentStep,
-    selectedSchema,
-    setSelectedSchema,
-    selectedOuId,
-    setSelectedOuId,
-    formValues,
-    setFormValues,
-    error,
-    setError,
-  } = useUserCreate();
-
-  const {data: userTypesData, isLoading: isUserTypesLoading, isFetching: isUserTypesFetching} = useGetUserTypes();
-  const {data: userTypeDetails, isLoading: isSchemaLoading} = useGetUserType(selectedSchema?.id);
-  const {
-    data: childOuData,
-    isLoading: isChildOuLoading,
-    error: childOuError,
-  } = useGetChildOrganizationUnits(selectedSchema?.ouId, {
-    limit: 1,
-    offset: 0,
-  });
-  const user = useThunderID().user as {ouId?: string} | null | undefined;
-  const tokenOuId = user?.ouId ?? null;
-  const isChildOuForbidden = (childOuError as {response?: {status?: number}} | null)?.response?.status === 403;
-  const isChildOuProbeFailed = !!childOuError && !isChildOuForbidden;
-  const userTypes = useMemo(() => userTypesData?.types ?? [], [userTypesData]);
-  const hasChildOUs = !isChildOuLoading && !childOuError && (childOuData?.totalResults ?? 0) > 0;
-  const isSingleUserType = !isUserTypesLoading && !isUserTypesFetching && userTypes.length === 1;
-
-  const activeSteps = useMemo((): UserCreateFlowStep[] => {
-    const base: UserCreateFlowStep[] = [];
-    if (!isSingleUserType) {
-      base.push(UserCreateFlowStep.USER_TYPE);
-    }
-    if (hasChildOUs) {
-      base.push(UserCreateFlowStep.ORGANIZATION_UNIT);
-    }
-    base.push(UserCreateFlowStep.USER_DETAILS);
-    return base;
-  }, [isSingleUserType, hasChildOUs]);
-
-  const steps: Partial<Record<UserCreateFlowStep, {label: string}>> = useMemo(() => {
-    const map: Partial<Record<UserCreateFlowStep, {label: string}>> = {};
-    if (!isSingleUserType) {
-      map.USER_TYPE = {label: t('users:createWizard.steps.userType')};
-    }
-    if (hasChildOUs) {
-      map.ORGANIZATION_UNIT = {label: t('users:createWizard.steps.organizationUnit')};
-    }
-    map.USER_DETAILS = {label: t('users:createWizard.steps.userDetails')};
-    return map;
-  }, [t, isSingleUserType, hasChildOUs]);
-
-  const [validationError, setValidationError] = useState<string | null>(null);
-  const [snackbarOpen, setSnackbarOpen] = useState(false);
-
-  const [stepReady, setStepReady] = useState<Record<UserCreateFlowStep, boolean>>({
-    USER_TYPE: false,
-    ORGANIZATION_UNIT: false,
-    USER_DETAILS: false,
-  });
-
-  const handleClose = (): void => {
-    if (createUserMutation.isPending) return;
-    Promise.resolve(navigate(routes.list())).catch((_error: unknown) => {
-      logger.error('Failed to navigate to users page', {error: _error});
+  const handleClose = useCallback(() => {
+    Promise.resolve(navigate(routes.list())).catch((err: unknown) => {
+      logger.error('Failed to navigate to users page', {error: err});
     });
-  };
+  }, [navigate, routes, logger]);
 
-  const handleBreadcrumbHome = (): void => {
-    if (createUserMutation.isPending) return;
-    Promise.resolve(navigate(routes.add())).catch((_error: unknown) => {
-      logger.error('Failed to navigate to add user page', {error: _error});
-    });
-  };
-
-  const handleStepReadyChange = useCallback((step: UserCreateFlowStep, isReady: boolean): void => {
-    setStepReady((prev) => ({
-      ...prev,
-      [step]: isReady,
-    }));
-  }, []);
-
-  const handleUserTypeStepReadyChange = useCallback(
-    (isReady: boolean): void => {
-      handleStepReadyChange(UserCreateFlowStep.USER_TYPE, isReady);
+  const handleStepLabelChange = useCallback(
+    (label: string) => {
+      setBreadcrumbs([t('users:addUser', 'Add User'), label]);
     },
-    [handleStepReadyChange],
+    [t],
   );
-
-  const handleOrganizationUnitStepReadyChange = useCallback(
-    (isReady: boolean): void => {
-      handleStepReadyChange(UserCreateFlowStep.ORGANIZATION_UNIT, isReady);
-    },
-    [handleStepReadyChange],
-  );
-
-  const handleUserDetailsStepReadyChange = useCallback(
-    (isReady: boolean): void => {
-      handleStepReadyChange(UserCreateFlowStep.USER_DETAILS, isReady);
-    },
-    [handleStepReadyChange],
-  );
-
-  const handleSchemaChange = useCallback(
-    (schema: typeof selectedSchema): void => {
-      if (schema?.id !== selectedSchema?.id) {
-        setFormValues({});
-        setSelectedOuId(null);
-        setStepReady((prev) => ({...prev, ORGANIZATION_UNIT: false, USER_DETAILS: false}));
-      }
-      setSelectedSchema(schema);
-    },
-    [selectedSchema, setSelectedSchema, setSelectedOuId, setFormValues],
-  );
-
-  // Resolve the OU for the selected type and move to the next step.
-  const advanceFromUserTypeStep = useCallback((): void => {
-    if (selectedSchema?.ouId && isChildOuLoading) {
-      // Wait for child OU probe to resolve before deciding
-      return;
-    }
-    if (isChildOuProbeFailed) {
-      setError(t('users:createWizard.errors.childOuProbeFailed'));
-      return;
-    }
-    if (hasChildOUs) {
-      setCurrentStep(UserCreateFlowStep.ORGANIZATION_UNIT);
-      return;
-    }
-    if (isChildOuForbidden) {
-      // User doesn't have permission to list child OUs — fall back to the OU from the access token
-      if (tokenOuId) {
-        setSelectedOuId(tokenOuId);
-        setCurrentStep(UserCreateFlowStep.USER_DETAILS);
-      } else {
-        setError(t('users:createWizard.errors.noOuAccess'));
-      }
-      return;
-    }
-    // No child OUs - skip OU step and use the schema's OU directly
-    setSelectedOuId(selectedSchema?.ouId ?? null);
-    setCurrentStep(UserCreateFlowStep.USER_DETAILS);
-  }, [
-    selectedSchema,
-    isChildOuLoading,
-    isChildOuProbeFailed,
-    hasChildOUs,
-    isChildOuForbidden,
-    tokenOuId,
-    setError,
-    setSelectedOuId,
-    setCurrentStep,
-    t,
-  ]);
-
-  // With only one user type, auto-select it and skip the selection step.
-  useEffect((): void => {
-    if (!isSingleUserType || currentStep !== UserCreateFlowStep.USER_TYPE) {
-      return;
-    }
-    if (!selectedSchema) {
-      setSelectedSchema(userTypes[0]);
-      return;
-    }
-    advanceFromUserTypeStep();
-  }, [isSingleUserType, currentStep, selectedSchema, userTypes, setSelectedSchema, advanceFromUserTypeStep]);
-
-  const handleSubmit = async (): Promise<void> => {
-    setValidationError(null);
-    setError(null);
-
-    if (!selectedSchema) {
-      setValidationError(t('users:createWizard.validationErrors.userTypeRequired'));
-      setSnackbarOpen(true);
-      return;
-    }
-
-    // Use the explicitly selected OU if available, otherwise fall back to the schema's OU
-    const ouId = selectedOuId ?? selectedSchema.ouId;
-    const trimmedOuId = ouId?.trim();
-    if (!trimmedOuId) {
-      setValidationError(t('users:createWizard.validationErrors.ouIdMissing'));
-      setSnackbarOpen(true);
-      return;
-    }
-
-    // Filter out empty/undefined attribute values to avoid sending
-    // blank fields that would fail backend schema validation.
-    const filteredAttributes = Object.fromEntries(
-      Object.entries(formValues).filter(([, v]) => v !== '' && v !== undefined && v !== null),
-    );
-
-    const requestBody = {
-      ouId: trimmedOuId,
-      type: selectedSchema.name,
-      attributes: filteredAttributes,
-    };
-
-    try {
-      await createUserMutation.mutateAsync(requestBody);
-      await navigate(routes.list());
-    } catch (submitError) {
-      logger.error('Failed to create user or navigate', {error: submitError});
-    }
-  };
-
-  const handleNextStep = (): void => {
-    switch (currentStep) {
-      case UserCreateFlowStep.USER_TYPE:
-        advanceFromUserTypeStep();
-        break;
-      case UserCreateFlowStep.ORGANIZATION_UNIT:
-        setCurrentStep(UserCreateFlowStep.USER_DETAILS);
-        break;
-      case UserCreateFlowStep.USER_DETAILS:
-        handleSubmit().catch(() => {
-          // Error handled in handleSubmit
-        });
-        break;
-      default:
-        break;
-    }
-  };
-
-  const handlePrevStep = (): void => {
-    switch (currentStep) {
-      case UserCreateFlowStep.ORGANIZATION_UNIT:
-        setCurrentStep(UserCreateFlowStep.USER_TYPE);
-        break;
-      case UserCreateFlowStep.USER_DETAILS:
-        if (hasChildOUs) {
-          setCurrentStep(UserCreateFlowStep.ORGANIZATION_UNIT);
-        } else {
-          setCurrentStep(UserCreateFlowStep.USER_TYPE);
-        }
-        break;
-      default:
-        break;
-    }
-  };
-
-  const renderStepContent = (): JSX.Element | null => {
-    switch (currentStep) {
-      case UserCreateFlowStep.USER_TYPE:
-        if ((isUserTypesLoading || isUserTypesFetching || isSingleUserType) && !error) {
-          // Show loading while resolving types; on auto-resolution failure, allow retry.
-          return (
-            <Box sx={{textAlign: 'center', py: 4}}>
-              <Typography variant="body2" color="text.secondary">
-                {t('common:status.loading')}
-              </Typography>
-            </Box>
-          );
-        }
-        return (
-          <ConfigureUserType
-            schemas={userTypes}
-            selectedSchema={selectedSchema}
-            onSchemaChange={handleSchemaChange}
-            onReadyChange={handleUserTypeStepReadyChange}
-          />
-        );
-      case UserCreateFlowStep.ORGANIZATION_UNIT:
-        if (!selectedSchema?.ouId) {
-          // Safety fallback — should not happen since the OU step is only shown
-          // when a schema with an ouId that has child OUs is selected.
-          setCurrentStep(UserCreateFlowStep.USER_TYPE);
-          return null;
-        }
-        return (
-          <ConfigureOrganizationUnit
-            key={selectedSchema.ouId}
-            rootOuId={selectedSchema.ouId}
-            selectedOuId={selectedOuId ?? ''}
-            onOuIdChange={setSelectedOuId}
-            onReadyChange={handleOrganizationUnitStepReadyChange}
-          />
-        );
-      case UserCreateFlowStep.USER_DETAILS:
-        if (isSchemaLoading) {
-          return (
-            <Box sx={{textAlign: 'center', py: 4}}>
-              <Typography variant="body2" color="text.secondary">
-                {t('common:status.loading')}
-              </Typography>
-            </Box>
-          );
-        }
-        if (!userTypeDetails) {
-          return null;
-        }
-        return (
-          <ConfigureUserDetails
-            key={selectedSchema?.id}
-            schema={userTypeDetails}
-            defaultValues={formValues}
-            onFormValuesChange={setFormValues}
-            onReadyChange={handleUserDetailsStepReadyChange}
-          />
-        );
-      default:
-        return null;
-    }
-  };
-
-  const getStepProgress = (): number => {
-    const currentIndex = activeSteps.indexOf(currentStep);
-    return ((currentIndex + 1) / activeSteps.length) * 100;
-  };
-
-  const getBreadcrumbSteps = (): UserCreateFlowStep[] => {
-    const currentIndex = activeSteps.indexOf(currentStep);
-    return activeSteps.slice(0, currentIndex + 1);
-  };
-
-  const handleBreadcrumbClick = useCallback(
-    (step: UserCreateFlowStep): void => {
-      setCurrentStep(step);
-    },
-    [setCurrentStep],
-  );
-
-  const handleCloseSnackbar = () => {
-    setSnackbarOpen(false);
-  };
-
-  const isLastStep = currentStep === activeSteps[activeSteps.length - 1];
 
   return (
     <Box sx={{minHeight: '100vh', display: 'flex', flexDirection: 'column'}}>
-      {/* Progress bar at the very top */}
-      <LinearProgress variant="determinate" value={getStepProgress()} sx={{height: 6}} />
+      <LinearProgress variant="determinate" value={0} sx={{height: 6}} />
 
       <Box sx={{flex: 1, display: 'flex', flexDirection: 'column'}}>
         {/* Header with close button and breadcrumb */}
         <Box sx={{p: 4, display: 'flex', justifyContent: 'space-between', alignItems: 'center'}}>
           <Stack direction="row" alignItems="center" spacing={2}>
             <IconButton
-              aria-label={t('common:actions.close')}
+              aria-label={t('common:actions.close', 'Close')}
               onClick={handleClose}
               sx={{
                 bgcolor: 'background.paper',
@@ -404,15 +299,10 @@ export default function UserCreatePage(): JSX.Element {
               <X size={24} />
             </IconButton>
             <AppBreadcrumbs
-              items={[
-                {key: 'add-user', label: t('users:addUser', 'Add User'), onClick: handleBreadcrumbHome},
-                {key: 'create-user', label: t('users:createWizard.title', 'Create User')},
-                ...getBreadcrumbSteps().map((step) => ({
-                  key: step,
-                  label: steps[step]?.label ?? step,
-                  onClick: () => handleBreadcrumbClick(step),
-                })),
-              ]}
+              items={breadcrumbs.map((label, idx) => ({
+                key: `breadcrumb-${idx}`,
+                label,
+              }))}
             />
           </Stack>
         </Box>
@@ -424,78 +314,32 @@ export default function UserCreatePage(): JSX.Element {
               flex: 1,
               display: 'flex',
               flexDirection: 'column',
-              py: 8,
+              pt: 8,
+              pb: 8,
               px: 20,
-              mx: currentStep !== UserCreateFlowStep.USER_DETAILS ? 'auto' : 0,
+              mx: 'auto',
               alignItems: 'flex-start',
+              justifyContent: 'flex-start',
             }}
           >
             <Box
               sx={{
                 width: '100%',
                 maxWidth: 800,
+                flex: 1,
                 display: 'flex',
                 flexDirection: 'column',
               }}
             >
-              {/* Error Alerts */}
-              {error && (
-                <Alert severity="error" sx={{my: 3}} onClose={() => setError(null)}>
-                  {error}
-                </Alert>
-              )}
-
-              {createUserMutation.error && (
-                <Alert severity="error" sx={{mb: 3}}>
-                  <Typography variant="body2" sx={{fontWeight: 'bold', mb: 0.5}}>
-                    {createUserMutation.error.message}
-                  </Typography>
-                </Alert>
-              )}
-
-              {renderStepContent()}
-
-              {/* Navigation buttons */}
-              <Stack direction="row" justifyContent="flex-end" alignItems="center" spacing={2} sx={{mt: 4}}>
-                {currentStep !== activeSteps[0] && (
-                  <Button variant="text" onClick={handlePrevStep} disabled={createUserMutation.isPending}>
-                    {t('common:actions.back')}
-                  </Button>
+              <InviteUser>
+                {(renderProps: InviteUserRenderProps) => (
+                  <UserCreateStepContent renderProps={renderProps} onStepLabelChange={handleStepLabelChange} />
                 )}
-
-                <Button
-                  variant="contained"
-                  disabled={
-                    !stepReady[currentStep] ||
-                    createUserMutation.isPending ||
-                    (currentStep === UserCreateFlowStep.USER_TYPE && Boolean(selectedSchema?.ouId) && isChildOuLoading)
-                  }
-                  sx={{minWidth: 140}}
-                  onClick={handleNextStep}
-                >
-                  {(() => {
-                    if (!isLastStep) return t('common:actions.continue');
-                    if (createUserMutation.isPending) return t('common:status.saving');
-                    return t('users:createUser.title');
-                  })()}
-                </Button>
-              </Stack>
+              </InviteUser>
             </Box>
           </Box>
         </Box>
       </Box>
-
-      {/* Validation Error Snackbar */}
-      <Snackbar
-        open={snackbarOpen}
-        autoHideDuration={6000}
-        onClose={handleCloseSnackbar}
-        anchorOrigin={{vertical: 'top', horizontal: 'right'}}
-      >
-        <Alert onClose={handleCloseSnackbar} severity="error" sx={{width: '100%'}}>
-          {validationError}
-        </Alert>
-      </Snackbar>
     </Box>
   );
 }
