+++ frontend/packages/design/src/components/flow/adapters/TextAdapter.tsx
@@ -22,6 +22,7 @@ import type {JSX} from 'react';
 import {useTranslation} from 'react-i18next';
 import useDesign from '../../../contexts/Design/useDesign';
 import type {FlowComponent} from '../../../models/flow';
+import mapEmbeddedFlowTextColor from '../../../utils/mapEmbeddedFlowTextColor';
 import {mapEmbeddedFlowTextVariant} from '../../../utils/mapEmbeddedFlowTextVariant';
 
 interface TextAdapterProps {
@@ -35,12 +36,14 @@ export default function TextAdapter({component, resolve}: TextAdapterProps): JSX
   const typographyVariant = mapEmbeddedFlowTextVariant(component.variant);
 
   const textAlign = component.align ?? (isDesignEnabled ? 'center' : 'left');
+  const color = mapEmbeddedFlowTextColor(component.color);
 
   return (
     <Typography
       id={component.id}
       className={[cn('Flow--text', `Text--${typographyVariant}`), component.classes].filter(Boolean).join(' ')}
       variant={typographyVariant}
+      color={color}
       sx={{mb: 1, textAlign}}
     >
       {t(resolve(component.label)!)}
