+++ frontend/packages/test-utils/src/setup.ts
@@ -161,6 +161,22 @@ vi.mock('@thunderid/react', async (importOriginal) => {
       isLoading: false,
     })),
     ThunderIDProvider: ({children}: {children: React.ReactNode}) => children,
+    InviteUser: ({children}: {children: (props: unknown) => React.ReactNode}) =>
+      children({
+        components: [],
+        values: {},
+        fieldErrors: {},
+        touched: {},
+        error: null,
+        isLoading: false,
+        handleInputChange: vi.fn(),
+        handleInputBlur: vi.fn(),
+        handleSubmit: vi.fn().mockResolvedValue(undefined),
+        resetFlow: vi.fn(),
+        isValid: false,
+        meta: null,
+        additionalData: undefined,
+      }),
   };
 });
